+++ rentplace/src/main/java/kattsyn/dev/rentplace/dtos/requests/RegisterRequest.java
@@ -1,4 +1,4 @@
-package kattsyn.dev.rentplace.dtos;
+package kattsyn.dev.rentplace.dtos.requests;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.AllArgsConstructor;
